Added r8 to pblh
*********1*********2*********3*********4*********5*********6*********7**
Longer commit message body describing the commit.
Can contain lists as follows:
	* Item 1
	* Item 2
	* Item 3

A good commit message should be written like an email, a subject
followed by a blank line, followed by a more descriptive body.

Can also contain a tag at the bottom describing what type of commit this is.
[BFB] - Bit-For-Bit
[FCC] - Flag Climate Changing
[Non-BFB] - Non Bit-For-Bit
[CC] - Climate Changing
[NML] - Namelist Changing

See confluence for a more detailed description about these tags.

+++ components/eam/src/physics/cam/gw_common.F90
@@ -902,8 +902,8 @@ end subroutine gw_oro_interface
 !==========================================================================
 function pblh_get_level_idx(height_array,pblheight)
 implicit none
-real(8),intent(in),dimension(pver) :: height_array
-real(8),intent(in) :: pblheight
+real(r8),intent(in),dimension(pver) :: height_array
+real(r8),intent(in) :: pblheight
 integer :: pblh_get_level_idx
        
 !local
